tests/test_error_handling.py
"""Test suite for error handling scenarios."""

from unittest.mock import MagicMock, patch

import numpy as np
import pytest
import torch
from PIL import Image

from src.components.user_interface import FeedbackUI, ProgressUI
from src.utils.core.error_handler import ErrorHandler


@pytest.fixture
def error_handler():
    """Create error handler instance."""
    return ErrorHandler()


@pytest.fixture
def mock_streamlit():
    """Mock streamlit components."""
    with patch("src.components.user_interface.st") as mock_st:
        mock_st.error = MagicMock()
        mock_st.warning = MagicMock()
        mock_st.info = MagicMock()
        yield mock_st


def test_error_handling_invalid_image(error_handler, mock_streamlit):
    """Test handling of invalid image input."""
    # Simulate invalid image error
    error = ValueError("Invalid image format")
    context = {"file_type": "invalid", "size": 0}

    error_handler.handle_error(error, context)

    # Verify error was logged
    assert len(error_handler.error_history) == 1
    assert error_handler.error_history[0]["type"] == "ValueError"
    assert "Invalid image format" in error_handler.error_history[0]["message"]


def test_error_handling_gpu_memory(error_handler, mock_streamlit):
    """Test handling of GPU memory errors."""
    # Simulate GPU memory error
    error = torch.cuda.OutOfMemoryError("CUDA out of memory")
    context = {"allocated_memory": "8GB", "required_memory": "12GB"}

    error_handler.handle_error(error, context)

    # Verify error was logged
    assert len(error_handler.error_history) == 1
    assert error_handler.error_history[0]["type"] == "OutOfMemoryError"
    assert "CUDA out of memory" in error_handler.error_history[0]["message"]


def test_error_handling_invalid_parameters(error_handler, mock_streamlit):
    """Test handling of invalid parameter errors."""
    # Simulate invalid parameter error
    error = ValueError("Invalid enhancement parameters")
    context = {"parameter": "sharpness", "value": -1.0}

    error_handler.handle_error(error, context)

    # Verify error was logged
    assert len(error_handler.error_history) == 1
    assert error_handler.error_history[0]["type"] == "ValueError"
    assert "Invalid enhancement parameters" in error_handler.error_history[0]["message"]


def test_error_handling_processing_timeout(error_handler, mock_streamlit):
    """Test handling of processing timeout errors."""
    # Simulate timeout error
    error = TimeoutError("Processing timeout")
    context = {"operation": "enhancement", "timeout": 30}

    error_handler.handle_error(error, context)

    # Verify error was logged
    assert len(error_handler.error_history) == 1
    assert error_handler.error_history[0]["type"] == "TimeoutError"
    assert "Processing timeout" in error_handler.error_history[0]["message"]


def test_error_handling_recovery(error_handler, mock_streamlit):
    """Test error recovery mechanisms."""
    # Simulate recoverable error
    error = RuntimeError("Temporary processing error")
    context = {"operation": "enhancement", "attempt": 1}

    error_handler.handle_error(error, context)

    # Verify error was logged
    assert len(error_handler.error_history) == 1
    assert error_handler.error_history[0]["type"] == "RuntimeError"

    # Verify tensor validation
    test_tensor = torch.randn(3, 64, 64)
    assert error_handler.verify_tensor(test_tensor)

    # Verify invalid tensor detection
    invalid_tensor = torch.tensor([float("nan")])
    assert not error_handler.verify_tensor(invalid_tensor)


def test_error_handling_feedback(mock_streamlit):
    """Test error handling in feedback collection."""
    ui = FeedbackUI()

    # Test with invalid parameters
    feedback = ui.collect_feedback(None, {"invalid": -1})

    # Verify feedback structure is valid despite invalid input
    assert isinstance(feedback, dict)
    assert "quality_rating" in feedback
    assert "aspect_ratings" in feedback
    assert "issues" in feedback
    assert "timestamp" in feedback


def test_error_handling_progress(mock_streamlit):
    """Test error handling in progress updates."""
    ui = ProgressUI()

    # Test with invalid metrics
    ui.update_progress({}, 0.0, {}, "Processing")

    # Verify progress handling of invalid data
    assert len(ui.metrics_history) == 1
    assert len(ui.quality_history) == 1
    assert len(ui.performance_history) == 1


def test_error_history_management(error_handler):
    """Test error history management."""
    # Add multiple errors
    errors = [
        (ValueError("Error 1"), {"context": "1"}),
        (RuntimeError("Error 2"), {"context": "2"}),
        (TypeError("Error 3"), {"context": "3"}),
    ]

    for error, context in errors:
        error_handler.handle_error(error, context)

    # Verify error history
    assert len(error_handler.error_history) == 3
    assert error_handler.error_history[0]["type"] == "ValueError"
    assert error_handler.error_history[1]["type"] == "RuntimeError"
    assert error_handler.error_history[2]["type"] == "TypeError"

    # Clear error history
    error_handler.clear_error_history()
    assert len(error_handler.error_history) == 0
