requirements.txt
altair==5.5.0
attrs==24.2.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.8
filelock==3.16.1
flake8==7.1.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imageio==2.36.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.13.0
mypy-extensions==1.0.0
narwhals==1.15.2
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.29.1
pyarrow==18.1.0
pycodestyle==2.12.1
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.24.0
scipy==1.14.1
six==1.17.0
smmap==5.0.1
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tifffile==2024.9.20
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
